+++ app/src/main/java/vola/systers/com/android/fragments/EventsMapFragment.java
@@ -1,4 +1,4 @@
-package vola.systers.com.volunteers_android.fragments;
+package vola.systers.com.android.fragments;
 
 
 
@@ -15,7 +15,7 @@ import com.google.android.gms.maps.SupportMapFragment;
 import com.google.android.gms.maps.model.LatLng;
 import com.google.android.gms.maps.model.MarkerOptions;
 
-import vola.systers.com.volunteers_android.R;
+import vola.systers.com.android.R;
 
 /*
  * @author divyapandilla
