+++ app/src/main/java/vola/systers/com/android/fragments/RequestsFragment.java
@@ -1,4 +1,4 @@
-package vola.systers.com.volunteers_android.fragments;
+package vola.systers.com.android.fragments;
 
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
@@ -6,17 +6,17 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-import vola.systers.com.volunteers_android.R;
+import vola.systers.com.android.R;
 
 /*
  * @author divyapandilla
  * @since 2017-06-11
  */
 
-public class StarredEventsFragment extends Fragment {
+public class RequestsFragment extends Fragment {
 
 
-    public StarredEventsFragment(){}
+    public RequestsFragment(){}
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
