+++ app/src/main/java/vola/systers/com/android/model/NavDrawerItem.java
@@ -1,4 +1,4 @@
-package vola.systers.com.volunteers_android.model;
+package vola.systers.com.android.model;
 
 /*
  * @author divyapandilla
