+++ app/src/main/res/layout/app_bar_main.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context="vola.systers.com.volunteers_android.activities.Menu">
+    tools:context="vola.systers.com.android.activities.Menu">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
